DAY-32/back/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
requests==2.32.3
python-dotenv==1.0.1
